init REST for question, user

+++ server/src/foundation/core/datalayer/__init__.py
@@ -2,31 +2,32 @@ from pymongo import MongoClient
 from bson import ObjectId
 from foundation.common.log import getLogger
 from flask import session
-
+import logging
 logger = getLogger(__name__)
 
 
 class MongoInterface(object):
     # client = MongoClient('mongodb://localhost:27017/')
     def __init__(self):
-        self.client = MongoClient(
-            'mongodb://data:chritsmasgood12@ds143971.mlab.com:43971/nvphu1306')
+        # self.client = MongoClient(
+        #     'mongodb://data:chritsmasgood12@ds143971.mlab.com:43971/nvphu1306')
+        self.client = MongoClient()
 
     @property
-    def db(self, dbname="nvphu1306"):
+    def db(self, dbname="test_module"):
         return self.client[dbname]
 
     def datasource(self, resource, query=None):
         source = self.db[resource]
         if query is None:
             query = {}
-        if session.get("AUTH_FIELD"):
-            query["user"] = session.get("username")
+        if session.get("AUTH_FIELD") and resource != "user":
+            query["userId"] = session.get("username")
         return (source, query)
 
     def find_one(self, resource, _id):
-        source, query = self.datasource(resource, {'_id': ObjectId(_id)})
-        data = source.find_one(query)
+        # source, query = self.datasource(resource, {'_id': ObjectId(_id)})
+        data = self.db[resource].find_one({'_id': ObjectId(_id)})
         return data
 
     def find(self, resource, query=None):
@@ -43,9 +44,49 @@ class MongoInterface(object):
             _id = ObjectId(_id)
         source, query = self.datasource(resource, {'_id': _id})
         data = source.update_one(query, update, **kwargs)
+        logging.warn("update by another user %s" % data)
         return data
 
     def delete_one(self, resource, _id):
         source, query = self.datasource(resource, {'_id': ObjectId(_id)})
+        if source.find_one(query) is None:
+            return False
         resp = source.delete_one(query)
-        return resp.acknowledged
+        return True
+
+    def check(self, resource, query=None):
+        if self.db[resource].find_one(query) is not None:
+            return True
+        else:
+            return False
+
+    def find_embedded(self, source, dist, _id, localField, foreignField):
+        h = self.db[source].aggregate([
+            {
+                "$lookup":
+                {
+                    "from": dist,
+                    "localField": localField,
+                    "foreignField": foreignField,
+                    "as": dist
+                }
+            },
+            {"$match": {"_id": ObjectId(_id)}},
+        ])
+        return h
+
+    def find_embedded_user(self, source, dist, _id, localField, foreignField):
+        h = self.db[source].aggregate([
+            {
+                "$lookup":
+                {
+                    "from": dist,
+                    "localField": localField,
+                    "foreignField": foreignField,
+                    "as": dist
+                }
+            },
+            {"$match": {"username": _id}},
+            {"$project": {"password": 0}}
+        ])
+        return h
